+++ pipescope/__init__.py
@@ -1,3 +1,3 @@
 """PipeScope: universal static analyzer for data pipelines."""
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"